requirements.txt
streamlit
easyocr
python-docx
pandas
Pillow
dateparser
numpy
opencv-python-headless
